Proposed .gitignore improvement
This will tell Git to track the `.dat` files on the USER partition of ArSim, so that someone working in simulation will know when they've made changes in TrakMaster. They won't know what the changes are, but it will clue them in that they need to save the `.dat`s to the Logical View.